status: introduce support for detailed status
Introduce a -detail flag, which is validated as a mode.

Enhance outputter to include Detail* methods analogous to the Basic*
methods, but including more information.

Begin by having all Detail* methods delegate to the corresponding
Basic* method.

Override that delegation behavior for tableOutput, and include the
project source information in the detailed output.

Use detailed output when the -detail flag is set.